integration test where it belongs
